Hazelcast log error fixed
