+++ devicehive-backend/src/main/java/com/devicehive/service/HazelcastService.java
@@ -20,10 +20,10 @@ package com.devicehive.service;
  * #L%
  */
 
-import com.devicehive.entity.HazelcastEntity;
-import com.devicehive.entity.HazelcastEntityComparator;
 import com.devicehive.model.DeviceCommand;
 import com.devicehive.model.DeviceNotification;
+import com.devicehive.model.HazelcastEntity;
+import com.devicehive.model.HazelcastEntityComparator;
 import com.devicehive.service.helpers.HazelcastHelper;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.IMap;
