+++ devicehive-common/src/main/java/com/devicehive/model/HazelcastEntity.java
@@ -1,4 +1,4 @@
-package com.devicehive.entity;
+package com.devicehive.model;
 
 /*
  * #%L
@@ -20,19 +20,11 @@ package com.devicehive.entity;
  * #L%
  */
 
-import java.io.Serializable;
-import java.util.Comparator;
 import java.util.Date;
-import java.util.Map;
 
-public class HazelcastEntityComparator implements Comparator<Map.Entry>, Serializable {
-    private static final long serialVersionUID = 5413354955792888308L;
+public interface HazelcastEntity {
 
-    @Override
-    public int compare(Map.Entry o1, Map.Entry o2) {
-        final Date o1Time = ((HazelcastEntity) o1.getValue()).getTimestamp();
-        final Date o2Time = ((HazelcastEntity) o2.getValue()).getTimestamp();
+    String getHazelcastKey();
 
-        return o1Time.compareTo(o2Time);
-    }
+    Date getTimestamp();
 }
