+++ devicehive-common/src/main/java/com/devicehive/model/HazelcastEntityComparator.java
@@ -1,4 +1,4 @@
-package com.devicehive.entity;
+package com.devicehive.model;
 
 /*
  * #%L
@@ -20,11 +20,19 @@ package com.devicehive.entity;
  * #L%
  */
 
+import java.io.Serializable;
+import java.util.Comparator;
 import java.util.Date;
+import java.util.Map;
 
-public interface HazelcastEntity {
+public class HazelcastEntityComparator implements Comparator<Map.Entry>, Serializable {
+    private static final long serialVersionUID = 5413354955792888308L;
 
-    String getHazelcastKey();
+    @Override
+    public int compare(Map.Entry o1, Map.Entry o2) {
+        final Date o1Time = ((HazelcastEntity) o1.getValue()).getTimestamp();
+        final Date o2Time = ((HazelcastEntity) o2.getValue()).getTimestamp();
 
-    Date getTimestamp();
+        return o1Time.compareTo(o2Time);
+    }
 }
